requirements.txt
asgiref==3.2.3
Django==3.0.14
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.11.0
Markdown==3.3.4
pytz==2019.3
sqlparse==0.3.0
